# Conflicts:
#	Commons/Commons.iml
#	Commons/src/main/java/models/Boat.java
#	Commons/src/test/java/models/BoatTest.java
#	Mock/Mock.iml
#	Mock/src/main/java/mockDatafeed/BoatMocker.java
#	Mock/src/main/java/mockDatafeed/BoatUpdater.java
#	Visualizer/src/main/java/controllers/MainController.java
#	Visualizer/src/main/java/controllers/RaceViewController.java
#	Visualizer/src/main/resources/css/theme.css
#	Visualizer/src/main/resources/raceView.fxml
#	Visualizer/src/test/java/steps/BoatStoppedSteps.java

+++ Commons/src/main/java/models/Competitor.java
@@ -83,11 +83,11 @@ public interface Competitor {
 
     void changeHeading(boolean upwind, double angle);
 
-    int getHealthLevel();
+    double getHealthLevel();
 
     void updateHealth(int delta);
 
-    int getMaxHealth();
+    double getMaxHealth();
 
     void setMaxHealth(int health);
